refactor: delete ObfuscateOps pass, add WidenConvLayer pass and InsertSkip pass
