+++ lib/Dialect/Torch/Transforms/WidenConvLayer.cpp
@@ -150,82 +150,20 @@ static void widenConvLayer(MLIRContext *context, Operation *f) {
                                                     resultTensorType, dense);
 }
 
-static void insertSkip(MLIRContext *context, Operation *f) {
-  // this demo insert a skip for the second convolution
-
-  llvm::SmallPtrSet<mlir::Operation *, 16> opWorklist;
-  f->walk([&](mlir::Operation *op) {
-    if (llvm::dyn_cast<AtenConvolutionOp>(op)) {
-      opWorklist.insert(op);
-    }
-  });
-  auto it = opWorklist.begin();
-  it++;
-  AtenConvolutionOp convOp = llvm::dyn_cast<AtenConvolutionOp>(*it);
-  mlir::OpBuilder builder(convOp);
-  mlir::IRRewriter rewriter(builder);
-  Location loc = convOp.getLoc();
-
-  // create a new conv with zero kernel and bias, to make sure output is the
-  // same as input
-  Value oldKernel = convOp.getOperand(1);
-  Value oldBias = convOp.getOperand(2);
-  // kernel
-  // shape: (new channels, old channels, height, width)
-  auto shape = oldKernel.getType().cast<ValueTensorType>().getSizes().vec();
-  // todo: 不确定有没有理解错，是添加一个1x1卷积核的卷积层吗
-  shape[0] = shape[1];
-  shape[2] = shape[3] = 1;
-  int kernelSize = shape[0] * shape[1] * shape[2] * shape[3];
-  std::vector<float> zeroKernelVec(kernelSize, 0);
-  auto resultTensorType = ValueTensorType::get(context, llvm::ArrayRef(shape),
-                                               builder.getF32Type());
-  auto dense = mlir::DenseElementsAttr::get(
-      mlir::RankedTensorType::get(llvm::ArrayRef(shape), builder.getF32Type()),
-      llvm::ArrayRef(zeroKernelVec));
-  Value zeroKernel =
-      rewriter.create<ValueTensorLiteralOp>(loc, resultTensorType, dense);
-  // bias
-  shape.erase(shape.begin() + 1, shape.end());
-  std::vector<float> zeroBiasVec(shape[0], 0);
-  resultTensorType = ValueTensorType::get(context, llvm::ArrayRef(shape),
-                                          builder.getF32Type());
-  dense = mlir::DenseElementsAttr::get(
-      mlir::RankedTensorType::get(llvm::ArrayRef(shape), builder.getF32Type()),
-      llvm::ArrayRef(zeroBiasVec));
-  Value zeroBias =
-      rewriter.create<ValueTensorLiteralOp>(loc, resultTensorType, dense);
-  // conv
-  Value zeroConv = rewriter.create<AtenConvolutionOp>(
-      loc, convOp.getOperand(0).getType(), convOp.getOperand(0), zeroKernel,
-      zeroBias, convOp.getOperand(3), convOp.getOperand(4),
-      convOp.getOperand(5), convOp.getOperand(6), convOp.getOperand(7),
-      convOp.getOperand(8));
-  // add
-  Value float0 =
-      rewriter.create<ConstantFloatOp>(loc, builder.getF64FloatAttr(0));
-  Value skip = rewriter.create<AtenAddTensorOp>(
-      loc, zeroConv.getType(), convOp.getOperand(0), zeroConv, float0);
-  rewriter.replaceOpWithNewOp<AtenConvolutionOp>(
-      convOp, convOp.getType(), skip, oldKernel, oldBias, convOp.getOperand(3),
-      convOp.getOperand(4), convOp.getOperand(5), convOp.getOperand(6),
-      convOp.getOperand(7), convOp.getOperand(8));
-}
-
 namespace {
-class ObfuscateOpsPass : public ObfuscateOpsBase<ObfuscateOpsPass> {
+class WidenConvLayerPass : public WidenConvLayerBase<WidenConvLayerPass> {
 public:
-  ObfuscateOpsPass() = default;
+  WidenConvLayerPass() = default;
   void runOnOperation() override {
     MLIRContext *context = &getContext();
     auto f = getOperation();
-    //    widenConvLayer(context, f);
-    insertSkip(context, f);
+    widenConvLayer(context, f);
   }
 };
 } // namespace
 
 std::unique_ptr<OperationPass<func::FuncOp>>
-mlir::torch::Torch::createObfuscateOpsPass() {
-  return std::make_unique<ObfuscateOpsPass>();
+mlir::torch::Torch::createWidenConvLayerPass() {
+  return std::make_unique<WidenConvLayerPass>();
 }
+
